Refactor DiscoverPresenterFactory to use Kotlin Inject:
- Extract interface and default implementation; update imports and usages.
